requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2023.3
nest-asyncio==1.5.8 